requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-dbbackup==4.1.0
django-environ==0.11.2
django-redis==5.4.0
django-storages==1.14.2
django-tinymce==3.6.1
dropbox==11.36.2
idna==3.6
pillow==10.2.0
ply==3.11
psycopg2==2.9.9
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stone==3.3.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
gunicorn==21.2.0
django-crontab==0.7.1
